trace: Add `Span::enter_scoped`
Currently, the primary way to use a span is to use `.enter` and pass a
closure to be executed under the span. While that is convenient in many
settings, it also comes with two decently inconvenient drawbacks:

 - It breaks control flow statements like `return`, `?`, `break`, and
   `continue`
 - It require re-indenting a potentially large chunk of code if you wish
   it to appear under a span

This branch adds a new `enter_scoped` function to `tokio_trace::Span`
that allows entering a span until a scope guard is dropped, as in:
```rust
let guard = span.enter_scoped();

// code here is within the span

drop(guard);

// code here is no longer within the span
```

Internally, the `enter` function (which was already implemented using
a package-private scope guard) was reimplemented on top of
`enter_scoped`.

As an aside, we might want to consider renaming `enter_scoped` to
`enter` and `enter` to...something else, but I'm punting on that for now
since it would require either a breaking change or a painful series of
deprecations...

Closes #1705

Signed-off-by: Eliza Weisman <eliza@buoyant.io>